Josshua Balita Update on Progam.cs_Ideas
Created a class named GameBoard, and added a main Program to test. Try it out and the program runs completely fine but this is only for test cases. Will add more classes (Class Player, Class AI).

NOTE: THIS ARE ONLY IDEAS, MAIN PROGRAM IS UNDER CONNECT4FINALPROJECT!!!!

SUBJECT FOR COLLABORATION